28/10 tp entregado

+++ src/clasesDAO/ClienteDAO.java
@@ -7,7 +7,7 @@ public interface ClienteDAO {
 	
 List<Cliente> listar();
 void eliminar(Cliente u);
-Cliente encontrar (Long id);
+Cliente encontrar (int id);
 void guardar (Cliente u);
 boolean validarCliente(String usuario, String contrasenia);
 int obtenerID(String usuario, String contrasenia);
