+++ src/clasesDAO/PeliculaDAO.java
@@ -6,7 +6,7 @@ public interface PeliculaDAO {
 	
 List<Pelicula> listar();
 void eliminar(Pelicula p);
-Pelicula encontrar (Long id);
+Pelicula encontrar (int id);
 void guardar (Pelicula p);
 int obtenerID(Pelicula p);
 }
