+++ src/clasesDAO/ReseniaDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 public interface ReseniaDAO {
 	List<Resenia> listarNoAprobadas();
 	void eliminar(Resenia r);
-	Resenia encontrar (Long id);
+	Resenia encontrar (int id);
 	void guardar (Resenia r);
 	void aprobarResenia(Resenia r);
 }
